demo web app finalized
okay so the web app demo is done man.

put all the UI fixes i planned when this new series of  commits and
updates

also added the delete option to the CRUD page

so that is our web demo app improved and ready

this does not mean this is the final fix.

but, for now, i am happy with these fixes

wrapping up this web demo app updates for now